requirements.txt
google-genai 
python-dotenv 
uvicorn 
fastapi
websockets
jinja2
